backend/requirements.txt
# Backend FastAPI Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Procesamiento de imágenes
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4
scikit-image

# Machine Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.14.0
timm

# BasicSR para EDSR
basicsr==1.4.2

# Utilidades
requests==2.31.0

# Logging y configuración
python-dotenv==1.0.0
loguru==0.7.2

# Performance
psutil==5.9.6

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1